problem with api in some locations
sometimes new parameters would show up in certain locations, so depending on that it might shift the keyword selection, and it throws the if function in the nicer looking one off and ruins the output. this can be adjusted manually for any parameters you want, just input the surrounding keywords into the variable "FS".